requirements.txt
torch
transformers
python-dotenv
openai
langchain
